+++ chapter-05/final/money/currency_test.go
@@ -6,12 +6,12 @@ import (
 )
 
 func TestGetCurrency_EUR(t *testing.T) {
-	expected := currency{
+	expected := Currency{
 		code:      "EUR",
 		precision: 2,
 	}
 
-	got, err := getCurrency("EUR")
+	got, err := parseCurrency("EUR")
 	if err != nil {
 		t.Errorf("expected no error, got %s", err.Error())
 	}
@@ -22,7 +22,7 @@ func TestGetCurrency_EUR(t *testing.T) {
 }
 
 func TestGetCurrency_UnknownCurrency(t *testing.T) {
-	_, err := getCurrency("UNKNOWN")
+	_, err := parseCurrency("UNKNOWN")
 	if !errors.Is(err, errUnknownCurrency) {
 		t.Errorf("expected error %s, got %v", errUnknownCurrency, err)
 	}
