+++ chapter-05/final/money/decimal.go
@@ -49,7 +49,7 @@ func parseNumber(value string) (number, error) {
 
 // applyChangeRate returns a new number representing n multiplied by the rate.
 // The precision is the same in and out.
-func (n number) applyChangeRate(rate changeRate, toPrecision int) number {
+func (n number) applyChangeRate(rate ChangeRate, toPrecision int) number {
 	converted := n.float() * float64(rate)
 
 	floor := math.Floor(converted)
@@ -69,9 +69,59 @@ func (n number) float() float64 {
 	return f
 }
 
+const (
+	// ErrInputTooSmall is returned if the amount to convert is too small, which could lead to precision issues.
+	ErrInputTooSmall = moneyError("input amount should be at least 1.00")
+	// ErrOutputTooSmall is returned if the converted amount is too small, in order to protect against precision issues.
+	ErrOutputTooSmall = moneyError("output amount is less than 1.00")
+	// ErrInputTooLarge is returned if the input amount is too large - this would cause floating point precision errors.
+	ErrInputTooLarge = moneyError("input amount over 10^15 is too large")
+	// ErrOutputTooLarge is returned if the converted amount is too large, to protect against floating point errors.
+	ErrOutputTooLarge = moneyError("output amount is too large (over 10^15)")
+
+	minAmount = 1
+	// maxAmount value is a thousand billion, using the short scale -- 10^12.
+	maxAmount = 1_000_000_000_000
+)
+
+func (n number) validateInput(sourceCurrency Currency) error {
+	switch {
+	case n.tooSmall():
+		return ErrInputTooSmall
+	case n.tooBig():
+		return ErrInputTooLarge
+	case n.tooPrecise(sourceCurrency):
+		return ErrOutputTooLarge
+	}
+	return nil
+}
+
+func (n number) tooPrecise(sourceCurrency Currency) bool {
+	return n.precision > sourceCurrency.precision
+}
+
+func (n number) validateOutput() error {
+	switch {
+	case n.tooSmall():
+		return ErrOutputTooSmall
+	case n.tooBig():
+		return ErrOutputTooLarge
+	}
+	return nil
+}
+
+func (n number) tooSmall() bool {
+	return n.integerPart < minAmount
+}
+
+func (n number) tooBig() bool {
+	return n.integerPart > maxAmount
+}
+
 // String implements stringer and returns the number formatted as
 // digits optionally a decimal point followed by digits.
 func (n number) String() string {
+	// for a precision of 2 digits formats %d.%02d
 	format := fmt.Sprintf("%%d.%%0%dd", n.precision)
 
 	return fmt.Sprintf(format, n.integerPart, n.decimalPart)
