--- chapter-05/final/money/statuscode.go
@@ -1,14 +1,0 @@
-package money
-
-import (
-	"fmt"
-	"net/http"
-)
-
-func checkStatusCode(statusCode int) error {
-	// TODO: Handle 4xx and 5xx separately ?
-	if statusCode != http.StatusOK {
-		return fmt.Errorf("invalid status code %d", statusCode)
-	}
-	return nil
-}
