--- chapter-05/layered/money/errors.go
@@ -1,9 +1,0 @@
-package money
-
-// moneyError defines a sentinel error.
-type moneyError string
-
-// moneyError implements the error interface.
-func (e moneyError) Error() string {
-	return string(e)
-}
